Fix ty type checker errors and enforce in CI
- bicubic_gen_payload.py: Use Image.Resampling.BICUBIC (modern Pillow API)
- bilinear_gen_payload.py: Add explicit .astype(np.float32) after cv2.cvtColor
  for type checker awareness (no runtime change - verified bit-for-bit)
- nearest_gen_payload.py: Remove Pillow <9.1 compat shim (requires Pillow ≥10)
- sanitizer.py: Add @overload signatures for sanitize_numeric return type
- app.py: Reformat long docstring to comply with line length limit
- pyproject.toml: Remove E501 ignore (all lines now under 100 chars)
- ci.yml: Remove continue-on-error to enforce type checking

Adds regression tests to ensure these assumptions hold in future versions.

+++ backend/adversarial_generators/bilinear_gen_payload.py
@@ -186,9 +186,9 @@ def main() -> None:
     decoy_bgr = cv2.imread(args.decoy, cv2.IMREAD_COLOR).astype(np.float32)
     target_bgr = cv2.imread(args.target, cv2.IMREAD_COLOR).astype(np.float32)
 
-    # Convert BGR to RGB
-    decoy_srgb = cv2.cvtColor(decoy_bgr, cv2.COLOR_BGR2RGB)
-    target_srgb = cv2.cvtColor(target_bgr, cv2.COLOR_BGR2RGB)
+    # Convert BGR to RGB (astype ensures type checker knows these are float32)
+    decoy_srgb = cv2.cvtColor(decoy_bgr, cv2.COLOR_BGR2RGB).astype(np.float32)
+    target_srgb = cv2.cvtColor(target_bgr, cv2.COLOR_BGR2RGB).astype(np.float32)
 
     # Convert to linear space
     decoy_lin = srgb2lin(decoy_srgb)
